scripts/librwtool/requirements.txt
argparse
pyelftools
pyyaml
capstone
keystone-engine
